requirements.txt
openai
langchain
streamlit
datasets
python-dotenv
langchain-community
tiktoken
chromadb